pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lookout"
version = "0.1.0"
description = "Lookout: Intelligent code pattern enforcement and discovery tool"
readme = "README.md"
license = {text = "Elastic License 2.0 (ELv2)"}
requires-python = ">=3.10"
authors = [
    { name = "Corey Taylor", email = "coreyt@users.noreply.github.com" }
]
keywords = ["linting", "semgrep", "pattern-discovery", "code-quality", "enforcement"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "semgrep>=1.50.0",
    "rich>=13.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "types-PyYAML",
    "types-beautifulsoup4",
]

[project.scripts]
lookout = "lookout.cli:cli"

[project.urls]
Homepage = "https://github.com/coreyt/lookout"
Repository = "https://github.com/coreyt/lookout"
Issues = "https://github.com/coreyt/lookout/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/lookout"]

[tool.hatch.build]
include = [
    "src/lookout/**/*.py",
    "src/lookout/**/*.yaml",
]

[tool.uv.workspace]
members = ["tui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/lookout"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 75

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]
